Fix for huge perf problem.
We have several queries for objects related to maps (zones, perimeters,
etc.). We we tell django to `select_related('map')`, we get a huge
performance loss because django fetches the maps' huge polygons and
converts them to python, even when we don't use them.

Using the `defer` functions works well to fix that problem.